Scenario: Race
The current Scenario Outline also asserts the roll history, which would
mean a bigger implementation step.  To work around this, we add a
separate, simpler scenario to slowly implement the Race scenarios.

The simplest implementation to pass all the Then steps was... to change
the /api/races mock from position 0 to 1 (!).